dev_ops_central_service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Marshmallow==0.15.0
flask-smorest==0.42.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0

# MySQL数据库驱动
PyMySQL==1.1.0
cryptography==41.0.7
mysql-connector-python==8.2.0

# 数据库连接池和性能优化 - 固定兼容版本
SQLAlchemy==1.4.53
alembic==1.12.1
